Add last OP codes

+++ NESEmulator.CPU/OPCodes/TSX_TransferStackPointerToX.cs
@@ -6,7 +6,7 @@ public class TSX_TransferStackPointerToX : IOPCode
 
     public bool Execute(CPU6502 cpu)
     {
-        cpu.X = cpu.StackPointer;
+        cpu.X = (byte)cpu.StackPointer;
 
         cpu.SetStatusFlag(CPUFlag.Z, cpu.X == 0x0000);
         cpu.SetStatusFlag(CPUFlag.N, (cpu.X & 0x80) > 0);
